+++ src/main/java/com/hexagonal/courseregistration/user/application/RegisterUserUseCase.java
@@ -1,7 +1,5 @@
 package com.hexagonal.courseregistration.user.application;
 
-import com.hexagonal.courseregistration.user.port.CheckExistUserPort;
-import com.hexagonal.courseregistration.user.port.SaveUserPort;
 import lombok.RequiredArgsConstructor;
 
 import static com.hexagonal.courseregistration.user.application.Message.ALREADY_EXIST_USER;
